Move scarb metadata call to binary
commit-id:de98a44d
