+++ crates/cairo-coverage-core/Cargo.toml
@@ -14,7 +14,6 @@ itertools.workspace = true
 ignore.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-scarb-metadata.workspace = true
 regex.workspace = true
 indoc.workspace = true
 rayon.workspace = true
