+++ crates/cairo-coverage-core/src/args.rs
@@ -1,13 +1,8 @@
-use camino::Utf8PathBuf;
-
 /// Options accepted by `cairo_coverage_core` `run` function.
 #[derive(Default, Clone)]
 pub struct RunOptions {
     /// Include additional components in the coverage report.
     pub include: Vec<IncludedComponent>,
-
-    /// Path to the project directory. If not provided, the project directory is inferred from the trace.
-    pub project_path: Option<Utf8PathBuf>,
 }
 
 /// Additional components that can be included in the coverage report.
