+++ crates/cairo-coverage/Cargo.toml
@@ -8,6 +8,7 @@ cairo-coverage-core = { path = "../cairo-coverage-core" }
 console.workspace = true
 camino.workspace = true
 anyhow.workspace = true
+scarb-metadata.workspace = true
 clap.workspace = true
 walkdir.workspace = true
